refactor(settings): extract footer component for better maintainability
Move footer-related logic from Settings component to a dedicated Footer component to improve code organization and maintainability. This reduces the complexity of the Settings component and makes the codebase easier to manage.
